// src/types/Repository.ts
// src/types/Repository.ts

type Owner = {
    login: string;
    id: number;
    node_id: string;
    avatar_url: string;
    gravatar_id: string;
    url: string;
    html_url: string;
    followers_url: string;
    following_url: string;
    gists_url: string;
    starred_url: string;
    subscriptions_url: string;
    organizations_url: string;
    repos_url: string;
    events_url: string;
    received_events_url: string;
    type: string;
    site_admin: boolean;
  };


interface Repository {
    id: number;
    name: string;
    description: string;
    html_url: string;
    language: string;
    stargazers_count: number;
    forks_count: number;
    owner: string; // Add 'owner' property with the appropriate type
    created_at: string; // Add 'created_at' property with the appropriate type
    updated_at: string; // Add 'updated_at' property with the appropriate type
  }
  
  
  export default Repository;